pyproject.toml
[tool.poetry]
name = "ssh-keyservice-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.8"
cryptography = "^44.0.1"
fastapi-azure-auth = "^5.0.1"
valkey = "^6.1.0"
pydantic = "^2.10.6"
pydantic-settings = "^2.7.1"
cachetools = "^5.5.1"
azure-identity = "^1.20.0"
azure-keyvault-secrets = "^4.9.0"


[tool.poetry.group.dev.dependencies]
uvicorn = "^0.34.0"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
